Moved settings reducers to separate file in reducers folder and updated all files to import actions and reducers correctly
